Deduplicate blueprint editor code
